requirements.txt
fastapi
uvicorn
email-validator
sqlalchemy
alembic
pydantic-settings
pyjwt
passlib[bcrypt]
python-jose[cryptography]
python-multipart